requirements.txt
requests
apscheduler
pandas
python-dotenv